+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Ensure worker message listener used one time are removed after usage.
+
 ## [2.7.0] - 2023-09-19
 
 ### Fixed